CRUD BASE/BACKEND/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
flasgger
